backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.0
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
